Switched over to game based time
Time in Object and derived classes is
no longer based on SDL ticks, but is
instead set inside the physics handler.

+++ src/Player.h
@@ -42,7 +42,7 @@ class Player : public Object
 
 		//add a missile, auto sets the sprite path to the default sprite
 		void addMissile(string sprite_path, string name, bool makeSelected,
-					int amnt, int mxDamage, int velocity, int lifespan);
+					int amnt, int mxDamage, float velocity, float lifespan);
 
 		void addMissile(string sprite_path, string name, bool makeSelected,
 					int amnt, int mxDamage);
